test(client-core): merge two HttpTransport.test.js files
